+++ recipes/configs/llama2/7B_lora_dpo.yaml
@@ -70,7 +70,6 @@ loss:
 epochs: 1
 max_steps_per_epoch: 1000
 gradient_accumulation_steps: 8
-compile: False
 
 # Logging
 output_dir: /tmp/lora_dpo_output/
