+++ torchtune/modules/loss/__init__.py
@@ -4,7 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .dpo import DPOLoss, IPOLoss, RSOLoss
-from .simPO import SimPOLoss
+from .dpo import DPOLoss, IPOLoss, RSOLoss, SimPOLoss
 
 __all__ = ["DPOLoss", "RSOLoss", "IPOLoss", "SimPOLoss"]
